Régler les autorisations pour Promoteur

+++ Controllers/PromoteursController.cs
@@ -7,6 +7,8 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using ASimmo.Data;
 using ASimmo.Models;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 namespace ASimmo.Controllers
 {
@@ -22,8 +24,18 @@ namespace ASimmo.Controllers
         // GET: Promoteurs
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.Promoteurs.Include(p => p.Type).Include(p => p.User);
-            return View(await applicationDbContext.ToListAsync());
+            if (this.HttpContext.User.IsInRole("Admin"))
+            {
+                var applicationDbContext = _context.Promoteurs.Include(p => p.Type).Include(p => p.User);
+                return View(await applicationDbContext.ToListAsync());
+            }  
+            else if (this.HttpContext.User.IsInRole("Agent"))
+            {
+                var applicationDbContext = _context.Promoteurs.Include(p => p.Type).Include(p => p.User).Where(p => p.User.UserName == this.HttpContext.User.Identity.Name);
+                return View(await applicationDbContext.ToListAsync());
+            }
+
+            return new ForbidResult();
         }
 
         // GET: Promoteurs/Details/5
@@ -47,10 +59,11 @@ namespace ASimmo.Controllers
         }
 
         // GET: Promoteurs/Create
+        [Authorize(Roles = "Admin, Agent")]
         public IActionResult Create()
         {
-            ViewData["TypeId"] = new SelectList(_context.TypesPromoteurs, "TypePromoteurId", "TypePromoteurId");
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id");
+            ViewData["TypeId"] = new SelectList(_context.TypesPromoteurs, "TypePromoteurId", "Libelle");
+            ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName");
             return View();
         }
 
@@ -59,7 +72,8 @@ namespace ASimmo.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("PromoteurId,UserId,TypeId")] Promoteur promoteur)
+        [Authorize(Roles = "Admin, Agent")]
+        public async Task<IActionResult> Create([Bind("Libelle,PromoteurId,UserId,TypeId")] Promoteur promoteur)
         {
             if (ModelState.IsValid)
             {
@@ -67,27 +81,40 @@ namespace ASimmo.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["TypeId"] = new SelectList(_context.TypesPromoteurs, "TypePromoteurId", "TypePromoteurId", promoteur.TypeId);
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id", promoteur.UserId);
+            ViewData["TypeId"] = new SelectList(_context.TypesPromoteurs, "TypePromoteurId", "Libelle", promoteur.TypeId);
+            ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName", promoteur.UserId);
             return View(promoteur);
         }
 
         // GET: Promoteurs/Edit/5
+        [Authorize(Roles = "Admin, Agent")]
         public async Task<IActionResult> Edit(int? id)
         {
+            
             if (id == null)
             {
                 return NotFound();
             }
 
-            var promoteur = await _context.Promoteurs.FindAsync(id);
+            var promoteur = await _context.Promoteurs.Include(p => p.User).FirstOrDefaultAsync(m => m.PromoteurId == id);
+
             if (promoteur == null)
             {
                 return NotFound();
             }
-            ViewData["TypeId"] = new SelectList(_context.TypesPromoteurs, "TypePromoteurId", "TypePromoteurId", promoteur.TypeId);
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id", promoteur.UserId);
-            return View(promoteur);
+
+            if (this.HttpContext.User.Identity.IsAuthenticated)
+            {
+                if (this.HttpContext.User.IsInRole("Admin") || this.HttpContext.User.Identity.Name == promoteur.User.UserName)
+                {
+                    ViewData["TypeId"] = new SelectList(_context.TypesPromoteurs, "TypePromoteurId", "Libelle", promoteur.TypeId);
+                    ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName", promoteur.UserId);
+                    return View(promoteur);
+                }
+            }
+            
+            return new ForbidResult();
+
         }
 
         // POST: Promoteurs/Edit/5
@@ -95,39 +122,54 @@ namespace ASimmo.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("PromoteurId,UserId,TypeId")] Promoteur promoteur)
+        [Authorize(Roles = "Admin, Agent")]
+        public async Task<IActionResult> Edit(int id, [Bind("Libelle,PromoteurId,UserId,TypeId")] Promoteur promoteur)
         {
+            
             if (id != promoteur.PromoteurId)
             {
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            if (PromoteurExists(promoteur.PromoteurId))
             {
-                try
+                var userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+
+                if (ModelState.IsValid)
                 {
-                    _context.Update(promoteur);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!PromoteurExists(promoteur.PromoteurId))
+                    try
                     {
-                        return NotFound();
+                        if (this.HttpContext.User.IsInRole("Admin") || promoteur.UserId == userId )
+                        {
+                            _context.Update(promoteur);
+                            await _context.SaveChangesAsync();
+                        }
+                        else
+                        {
+                            return new ForbidResult();
+                        }
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
+
                         throw;
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
             }
-            ViewData["TypeId"] = new SelectList(_context.TypesPromoteurs, "TypePromoteurId", "TypePromoteurId", promoteur.TypeId);
-            ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id", promoteur.UserId);
+            else
+            {
+                return NotFound();
+            }
+            
+
+            ViewData["TypeId"] = new SelectList(_context.TypesPromoteurs, "TypePromoteurId", "Libelle", promoteur.TypeId);
+            ViewData["UserId"] = new SelectList(_context.Users, "Id", "UserName", promoteur.UserId);
             return View(promoteur);
         }
 
         // GET: Promoteurs/Delete/5
+        [Authorize(Roles = "Admin, Agent")]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -139,23 +181,41 @@ namespace ASimmo.Controllers
                 .Include(p => p.Type)
                 .Include(p => p.User)
                 .FirstOrDefaultAsync(m => m.PromoteurId == id);
+
             if (promoteur == null)
             {
                 return NotFound();
             }
-
-            return View(promoteur);
+            if (this.HttpContext.User.IsInRole("Admin") || this.HttpContext.User.Identity.Name == promoteur.User.UserName)
+            {
+                return View(promoteur);
+            }
+            else
+            {
+                return new ForbidResult();
+            }
         }
 
         // POST: Promoteurs/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin, Agent")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var userId = this.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value;
             var promoteur = await _context.Promoteurs.FindAsync(id);
-            _context.Promoteurs.Remove(promoteur);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+
+            if (this.HttpContext.User.IsInRole("Admin") || promoteur.UserId == userId)
+            {
+                _context.Promoteurs.Remove(promoteur);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                return new ForbidResult();
+            }
+
         }
 
         private bool PromoteurExists(int id)
